Implement editorconfig checker workflow
This implementation can detect bad trailing whitespace and missing LF
at the EOF. But some tabs indentation checker is still a bit conflict
with Linux kernel coding style when it lines up multiple arguments that
are broken down into multiple lines.

We plan to fix the conflict in the near future, but it is not an urgent
issue at the moment.

Fixes: https://github.com/GNUWeeb/GNUWeebBot/issues/30

Suggested-by: Sri Aspari <sriaspari@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>