Add silent kick and silent ban commands
Sometimes it's annoying to see a lot of ban and kick commands being
invoked to get rid of spammers. As such this issue, we introduce
silent version of these commands.

The silent version will perform the action silently and will delete
the command message sent by the invoker, including replied message.

Suggested-by: Arthur Lapz <arthurlapz@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>