modules/003_admin: Add error report for silent commands
When delete message fails, we don't know that it fails. Because
everything is silent. This can be confusing for debugging and user
won't be noticed what's wrong with the bot regarding the failure
deletion.

This patch adds a reply message to silent commands when invoked the
commands fail. The reply message will contain the reason of failure
from TG API.

Fixes: 93b20433d0aa90aa83e4e7015e281a4819ecdf9a ("Add silent kick and silent ban commands")
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>
Signed-off-by: Komori Kuzuyu <komori.kzy@gmail.com>